chore: move assemble responsibility to process
